ref(http-exc): Create a module exceptions for all http exceptions
To avoid the repetitive multi-lines of code, all http exceptions are
rewritten in a function with customize message based on the exact error /
exceptions.

+++ backend/app/api/routes/users.py
@@ -1,34 +1,41 @@
-from typing import Any
-
+# type: ignore
 import fastapi
 
 from app.api.dependencies.authorization import retrieve_current_user_auth
 from app.api.dependencies.database import get_repository
+from app.api.exceptions.http_exc_400 import http400_exc_bad_email_request, http400_exc_bad_username_request
+from app.api.exceptions.http_exc_404 import http404_exc_id_not_found
 from app.core.config import get_settings
-from app.core.settings.app import AppSettings
 from app.db.repositories.users import UsersRepository
 from app.models.domain.users import User
 from app.models.schemas.users import UserInResponse, UserInUpdate, UserWithToken
 from app.services.authentication import check_email_is_taken, check_username_is_taken
 from app.services.jwt import generate_access_token
 
-router = fastapi.APIRouter(prefix="/users", tags=["User"])
+router = fastapi.APIRouter(prefix="/user", tags=["users"])
+settings = get_settings()
 
 
 @router.get(
-    path="/user",
+    path="",
     name="users:get-current-user",
     response_model=UserInResponse,
     status_code=fastapi.status.HTTP_200_OK,
 )
 async def retrieve_current_user(
+    id: int,
     user: User = fastapi.Depends(retrieve_current_user_auth()),
-    settings: AppSettings = fastapi.Depends(get_settings),
 ) -> UserInResponse:
+
+    if id != user.id_:
+
+        return await http404_exc_id_not_found(id=id)
+
     token = generate_access_token(
         user,
         settings.secret_key,
     )
+
     return UserInResponse(
         user=UserWithToken(
             username=user.username,
@@ -45,24 +52,21 @@ async def retrieve_current_user(
     status_code=fastapi.status.HTTP_200_OK,
 )
 async def update_current_user(
+    id: int,
     user_update: UserInUpdate = fastapi.Body(..., embed=True, alias="user"),
     current_user: User = fastapi.Depends(retrieve_current_user_auth()),
     users_repo: UsersRepository = fastapi.Depends(get_repository(UsersRepository)),
-    settings: AppSettings = fastapi.Depends(get_settings),
 ) -> UserInResponse:
+    if id != current_user.id_:
+        return await http404_exc_id_not_found(id=id)
+
     if user_update.username and user_update.username != current_user.username:
         if await check_username_is_taken(users_repo, user_update.username):
-            raise fastapi.HTTPException(
-                status_code=fastapi.status.HTTP_400_BAD_REQUEST,
-                detail="Username is taken!",
-            )
+            return await http400_exc_bad_username_request(username=user_update.username)
 
     if user_update.email and user_update.email != current_user.email:
         if await check_email_is_taken(users_repo, user_update.email):
-            raise fastapi.HTTPException(
-                status_code=fastapi.status.HTTP_400_BAD_REQUEST,
-                detail="Email is taken!",
-            )
+            return await http400_exc_bad_email_request(email=user_update.email)
 
     user = await users_repo.update_user(user=current_user, **user_update.dict())
 
@@ -86,10 +90,17 @@ async def update_current_user(
     status_code=fastapi.status.HTTP_202_ACCEPTED,
 )
 async def delete_current_user(
+    id: int,
     current_user: User = fastapi.Depends(retrieve_current_user_auth()),
     users_repo: UsersRepository = fastapi.Depends(get_repository(UsersRepository)),
-) -> Any:
+) -> fastapi.responses.JSONResponse:
 
-    user = await users_repo.get_user_by_id(id=current_user.id_)  # type: ignore
+    if id != current_user.id_:
 
-    return await users_repo.delete_user(id=user.id_)  # type: ignore
+        return await http404_exc_id_not_found(id=id)
+
+    user = await users_repo.get_user_by_id(id=current_user.id_)
+
+    await users_repo.delete_user(id=user.id_)
+
+    return {"msg": f"User with ID {id} is successfully deleted!"}
